requirements.txt
mcp>=1.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
